PCT-635: s/uint8/uint/ since bool metrics can be up to query_count (Cnt) which is uint

+++ log/stats.go
@@ -36,7 +36,7 @@ type NumberStats struct {
 
 type BoolStats struct {
 	Cnt  uint
-	True uint8
+	True uint
 }
 
 func NewEventStats() *EventStats {
